+++ src/base/simplex.rs
@@ -8,7 +8,7 @@ use crate::{
 use core::simd::{f32x2, f32x4};
 
 /// 2/3/4 dimensional Simplex noise.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Simplex;
 
 impl_noise!(234 Simplex);